few changes
made current_time and its related identifiers to type 'long' due to implicit conversion warnings
rewrote 'print_turnaround_times()' for use an enhanced for loop

+++ header.h
@@ -7,12 +7,12 @@
 
 void readData();
 
-void add_process(int current_time);
+void add_process(long current_time);
 
-void extra_memory_helper(int current_time);
+void extra_memory_helper(long current_time);
 
-void remove_process(int current_time);
+void remove_process(long current_time);
 
 void print_turnaround_times();
 
-void assign_memory_for_process(int crrent_time);
+void assign_memory_for_process(long crrent_time);
